refactor:♻️FieldBatchInput型をimport機能に移動(#4)
