+++ internal/features/import/application/usecase/process_import_test.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/mktkhr/field-manager-api/internal/features/import/domain/entity"
-	"github.com/mktkhr/field-manager-api/internal/features/shared/types"
 )
 
 // mockStorageClient はStorageClientのモック実装
@@ -48,7 +47,7 @@ type mockFieldRepository struct {
 	err error
 }
 
-func (m *mockFieldRepository) UpsertBatch(ctx context.Context, inputs []types.FieldBatchInput) error {
+func (m *mockFieldRepository) UpsertBatch(ctx context.Context, inputs []FieldBatchInput) error {
 	return m.err
 }
